Restructured files
Moved files around and changed premake file (just so it looks a little neater)
