+++ src/imgui.ini
@@ -30,7 +30,7 @@ Collapsed=0
 DockId=0x00000003,0
 
 [Window][Test Window]
-Pos=661,1037
+Pos=2,599
 Size=854,480
 Collapsed=0
